+++ NEWS.md
@@ -42,7 +42,8 @@ No longer force render data as JSON if using an HTML template.
 - Allow customising the path to pattern converter.
 - Added `get_header` method to the retrieve a specific method.
 - Fix htmlwidget response.
-- Add `png` and `jpeg` responses.
+- Added `png` and `jpeg` methods to `Response` to serve images.
+- Added `plot_png` and `plot_jpeg` methods to `Response` to serve plots.
 
 # ambiorix 2.0.0